--- a/examples/simple.go
+++ b/examples/simple.go
@@ -9,8 +9,6 @@ import (
 
 const expr = "0 * 9 LW JAN-OCT 1-5 2000/10"
 
-// const expr = "00 11,16 * * *"
-
 func main() {
 	exprDesc, err := cron.NewDescriptor(
 		cron.Use24HourTimeFormat(true),