```
$ benchstat before_result.pprof after_result.pprof 
name                                old time/op    new time/op    delta
ExpressionDescriptor_ToDescription    15.6µs ± 3%    11.8µs ± 1%  -24.52%  (p=0.000 n=10+9)

name                                old alloc/op   new alloc/op   delta
ExpressionDescriptor_ToDescription    5.63kB ± 0%    4.59kB ± 0%  -18.47%  (p=0.000 n=10+10)

name                                old allocs/op  new allocs/op  delta
ExpressionDescriptor_ToDescription      92.0 ± 0%      90.0 ± 0%   -2.17%  (p=0.000 n=10+10)

```